update of seeds
