+++ seeds/index.js
@@ -1,5 +1,5 @@
 const sequelize = require('../config/connection');
-const gifts = require('./gifts');
+const gifts = require('./giftsData');
 
 const seedAll = async () => {
   await sequelize.sync({ force: true });
